Added panels comp to wrap visible panels for authenticated users
