+++ src/components/__tests__/ToolPanel.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ToolPanel, { ToolPanelProps } from '../ToolPanel'
 import { render, cleanup } from 'utils/test-utils'
-import { ToolbarPanelOptions } from '../App'
+import { ToolbarPanelOptions } from '../Panels'
 
 afterEach(() => cleanup())
