+++ src/utils/test-utils.js
@@ -6,19 +6,22 @@ import { StarProvider } from 'state/star-context'
 import { SearchProvider } from 'state/search-context'
 import { TagProvider } from 'state/tag-context'
 import { DndProvider } from 'react-dnd'
+import { GithubProvider } from 'state/github-context'
 import HTML5Backend from 'react-dnd-html5-backend'
 
 const GlobalProviders = ({ children }) => {
   return (
-    <ThemeProvider theme={theme}>
-      <DndProvider backend={HTML5Backend}>
-        <TagProvider>
-          <SearchProvider>
-            <StarProvider>{children}</StarProvider>
-          </SearchProvider>
-        </TagProvider>
-      </DndProvider>
-    </ThemeProvider>
+    <GithubProvider>
+      <ThemeProvider theme={theme}>
+        <DndProvider backend={HTML5Backend}>
+          <TagProvider>
+            <SearchProvider>
+              <StarProvider>{children}</StarProvider>
+            </SearchProvider>
+          </TagProvider>
+        </DndProvider>
+      </ThemeProvider>
+    </GithubProvider>
   )
 }
